Rename : 폴더이름 수정
- 오타 수정 assetes -> assets
